Updated BuildConfig to promote compatibility.
Since I develop in 2.1.1 for now, I've modified the BuildConfig so that
it does a few nifty things:

1. It sets up spock correctly based on Grails version, at least for
2.0/2.1 and 2.2 .

2. I've disabled the export of the dependency on spock resources, as that was
causing major problems and should not be included in applications anyway; that
is only for the plugin and should remain only in the plugin.

3. I've disabled the export of jaxrs. While I understand that may make
it convenient, I feel that it is more appropriate to allow the
application to specify what version of JaxRS is it is using. Since I am
running 2.1.1, JaxRS 0.9 is my compatible version. Otherwise, I would
have to specifically export jaxrs from this plugin's transitive
dependencies, just to include my version; this is not good plugin
practice.