Removed presumptuous hard-coding.
The field that was removed assumed that the application is going to use
JaxRS's default /api path, however the application should be allowed to
use whatever paths it so desires.

By removing it here and adding it in the annoation, not only does it
provide better flexibility for applications, but that is also how JaxRS
does it as well.

For example, if the config field is set to [ '/v1' ], now both the @Api
and @Path annoations for swagger and JaxRS respectively will look the
same.