Lean away from grails.serverURL .
According to the Grails 2.x documentation, the behavior of the Config
grails.serverURL has changed behavior with respect to environments.

This Config also appears more associated with createLink and redirect
methods.

I have changed the code to not use grails.serverURL at all, but rather
to use the grailsLinkGenerator bean within the controller.

Anomaly: Assuming your production server is "http://sudofu.me" and you
have a grails app that uses the grails.serverURL config for this plugin,
the behavior between grails run-app and deploying WAR changes.

In run-app, the serverURL is respected and the plugin appears to work
fine. However, when WAR'd, the
"${grailsApplication.config.grails.serverURL}/api-docs" value in the
showRestApi view actually produces the following:

http://sudofu.me/sudofu.me/api-docs

In other words, it appends the hostname an additional time, and
therefore the api-docs link is not readable.

By leaning away from the grails.serverURL, the behavior is consistent
between both run methods.

Caveat: The link generated by the grailsLinkGenerator bean has different
states. At first I had placed this in the
Swagger4jaxrsGrailsPlugin.groovy file, however that lead to the address
being:

http://localhost:8080/

Rather than the expected:

http://localhost:8080/MyApp/

As it turns out, the grailsLinkGenerator is not reliably fully
initialized until the application is running. Therefore, I had to
migrate this configuration for the swaggerConfig bean into the
showRestApi controller, so that the address is consistent across run
methods.

I also explicitly pass this root address to the index.gsp view, as that
is the cleanest way to do so.